improve tests and async resource management
