+++ packages/datadog-plugin-jest/test/jest-test.js
@@ -1,3 +1,5 @@
+const http = require('http')
+
 describe('jest-test-suite', () => {
   // eslint-disable-next-line
   jest.setTimeout(200)
@@ -23,6 +25,20 @@ describe('jest-test-suite', () => {
       done()
     }, 100)
   })
+  it('can do integration http', (done) => {
+    const req = http.request('http://test:123', (res) => {
+      expect(res.statusCode).toEqual(200)
+      done()
+    })
+    req.end()
+  })
+  // only run for jest-circus tests
+  // eslint-disable-next-line
+  if (jest.retryTimes) {
+    it.each([[1, 2, 3], [2, 3, 5]])('can do parameterized test', (a, b, expected) => {
+      expect(a + b).toEqual(expected)
+    })
+  }
   it('promise passes', () => {
     return new Promise((resolve) =>
       setTimeout(() => {
@@ -65,3 +81,16 @@ describe('jest-test-suite', () => {
   })
   it.todo('skips todo')
 })
+
+// only run for jest-circus tests
+// eslint-disable-next-line
+if (jest.retryTimes) {
+  describe('jest-circus-test-retry', () => {
+    // eslint-disable-next-line
+    jest.retryTimes(2)
+    let retryAttempt = 0
+    it('can retry', () => {
+      expect(retryAttempt++).toEqual(2)
+    })
+  })
+}
